+++ core/src/main/java/nl/nn/adapterframework/extensions/javascript/J2V8.java
@@ -28,10 +28,10 @@ import com.eclipsesource.v8.V8Object;
 
 import nl.nn.adapterframework.core.ISender;
 import nl.nn.adapterframework.core.PipeLineSession;
-import nl.nn.adapterframework.extensions.graphviz.ResultHandler;
 import nl.nn.adapterframework.stream.Message;
 import nl.nn.adapterframework.util.AppConstants;
 import nl.nn.adapterframework.util.LogUtil;
+import nl.nn.adapterframework.util.flow.graphviz.ResultHandler;
 
 public class J2V8 implements JavascriptEngine<V8> {
