+++ core/src/main/java/nl/nn/adapterframework/extensions/javascript/Rhino.java
@@ -15,10 +15,10 @@
 */
 package nl.nn.adapterframework.extensions.javascript;
 
-import nl.nn.adapterframework.extensions.graphviz.ResultHandler;
 import org.mozilla.javascript.*;
 
 import nl.nn.adapterframework.core.PipeLineSession;
+import nl.nn.adapterframework.util.flow.graphviz.ResultHandler;
 import nl.nn.adapterframework.core.ISender;
 
 public class Rhino implements JavascriptEngine<Context> {
