+++ core/src/main/java/nl/nn/adapterframework/util/flow/FlowDiagramAdapterWrapper.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package nl.nn.adapterframework.extensions.graphviz;
+package nl.nn.adapterframework.util.flow;
 
 import java.io.IOException;
 
@@ -22,7 +22,6 @@ import nl.nn.adapterframework.configuration.ConfigurationWarnings;
 import nl.nn.adapterframework.configuration.SuppressKeys;
 import nl.nn.adapterframework.core.Adapter;
 import nl.nn.adapterframework.util.AppConstants;
-import nl.nn.adapterframework.util.flow.FlowDiagramManager;
 
 public class FlowDiagramAdapterWrapper extends AdapterLifecycleWrapperBase {
 	private FlowDiagramManager flowDiagramManager;
