+++ core/src/main/java/nl/nn/adapterframework/util/flow/graphviz/GraphvizEngine.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package nl.nn.adapterframework.extensions.graphviz;
+package nl.nn.adapterframework.util.flow.graphviz;
 
 import java.io.IOException;
 import java.net.URL;
