+++ core/src/main/resources/SpringApplicationContext.xml
@@ -62,7 +62,7 @@
 
 	<bean id="adapterLifecycleWrapper" autowire="byName" scope="prototype" abstract="true" />
 
-	<bean name="adapterFlowProcessor" class="nl.nn.adapterframework.extensions.graphviz.FlowDiagramAdapterWrapper" parent="adapterLifecycleWrapper"/>
+	<bean name="adapterFlowProcessor" class="nl.nn.adapterframework.util.flow.FlowDiagramAdapterWrapper" parent="adapterLifecycleWrapper"/>
 	<bean name="adapterJmxProcessor" class="nl.nn.adapterframework.jmx.JmxAdapterWrapper" parent="adapterLifecycleWrapper"/>
 
 	<bean id="mbeanServer" class="org.springframework.jmx.support.MBeanServerFactoryBean">
