+++ core/src/test/java/nl/nn/adapterframework/extensions/javascript/NashornTest.java
@@ -15,10 +15,11 @@
 */
 package nl.nn.adapterframework.extensions.javascript;
 
-import nl.nn.adapterframework.extensions.graphviz.ResultHandler;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+
+import nl.nn.adapterframework.util.flow.graphviz.ResultHandler;
 
 public class NashornTest {
 	private Nashorn engine;
