+++ core/src/test/java/nl/nn/adapterframework/util/flow/graphviz/GraphvizEngineTest.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
  */
-package nl.nn.adapterframework.extensions.graphviz;
+package nl.nn.adapterframework.util.flow.graphviz;
 
 import static nl.nn.adapterframework.testutil.TestAssertions.assertEqualsIgnoreWhitespaces;
 import static org.junit.Assert.assertEquals;
